Fixed all compiler warnings. glm files still throw loads, but I'm not messing with other people's C code. There is one warning left, this is and added message of something that needs to be looked into in Texture.cpp

+++ WarpDrive/BaseSystem/Game.cpp
@@ -53,7 +53,7 @@ void Game::Draw()
 
 void Game::resetState(int i)
 {
-	if (i <0 || i >= states.size())
+	if (i < 0 || static_cast<size_t>(i) >= states.size())
 	{
 		return;
 	}
